Rename "namespace geometry_world" to "namespace scene_graph"
